requirements.txt
appnope==0.1.0
asn1crypto==0.23.0
bcrypt==3.1.4
boto3==1.4.7
botocore==1.7.36
CacheControl==0.11.7
cement==2.10.2
census==0.8.4
cffi==1.11.2
cryptography==2.1.1
decorator==4.1.2
Django==1.11.6
django-extensions==1.9.6
django-filter==1.1.0
django-postgres-copy==2.0.1
django-storages==1.6.5
django-uuslug==1.1.8
djangorestframework==3.7.1
docutils==0.14
elex==2.4.1
Fabric3==1.13.1.post1
geojson==2.3.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jellyfish==0.5.6
Jinja2==2.9.6
jmespath==0.9.3
lockfile==0.12.2
Markdown==2.6.9
MarkupSafe==1.0
nltk==3.2.5
olefile==0.44
paramiko==2.3.1
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pyasn1==0.3.7
pycparser==2.18
pydot==1.2.3
Pygments==2.2.0
pymongo==3.3.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyshp==1.2.12
python-dateutil==2.2
python-slugify==1.2.4
pytz==2017.2
requests==2.11.1
s3transfer==0.1.11
simplegeneric==0.8.1
six==1.11.0
tqdm==4.19.4
traitlets==4.3.2
ujson==1.35
Unidecode==0.4.21
us==1.0.0
wcwidth==0.1.7
